consolidate coverage tests into existing test suites and remove redundant test file
